requirements.txt
django==1.7.1
MySQL-python==1.2.5
django-decouple==2.1
django-ckeditor==4.2.8
easy_thumbnails==2.0.1
django-grappelli==2.6.5
django-image-cropping==0.8.2